+++ tests/utils.py
@@ -1,5 +1,6 @@
 import os
 import sys
+from textwrap import dedent
 
 from mock import patch
 from moban import file_system
@@ -54,3 +55,10 @@ class Docs(object):
     def run_moban_with_fs(self, moban_cli, working_directory, assertions):
         os.chdir(os.path.join(self.base_folder, working_directory))
         run_moban_with_fs(moban_cli, None, assertions)
+
+
+def custom_dedent(long_texts):
+    refined = dedent(long_texts)
+    if refined.startswith("\n"):
+        refined = refined[1:]
+    return refined
